Move and split of App to component and container
