+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import SciDataFormContainer from './views/containers/scidataForm';
 import { Provider } from 'react-redux';
 import { JsonFormsState } from '@jsonforms/core';
 import { materialCells, materialRenderers } from '@jsonforms/material-renderers';
@@ -9,18 +9,25 @@ import { materialCells, materialRenderers } from '@jsonforms/material-renderers'
 import configureStore from './state/store';
 
 
-const sciDataState: JsonFormsState = {
+const sciDataInitialState: JsonFormsState = {
   jsonforms: {
     cells: materialCells,
     renderers: materialRenderers
   }
 };
 
-const store = configureStore(sciDataState);
+/*
+const initialState = {
+  sciDataState: sciDataInitialState
+}
+*/
+
+const store = configureStore(sciDataInitialState);
+
 
 ReactDOM.render(
   <Provider store={store}>
-    <App />
+    <SciDataFormContainer />
   </Provider>,
   document.getElementById('root')
 );
