+++ src/state/store.js
@@ -1,4 +1,4 @@
-import { createStore, combineReducers } from "redux";
+import { createStore } from "redux";
 import { Actions } from '@jsonforms/core';
 
 //import * as reducers from "./ducks";
@@ -9,6 +9,7 @@ import scidataJsonSchema from './ducks/scidata/jsonSchema.json';
 import scidataUiSchema from './ducks/scidata/uiSchema.json';
 import { default as scidataJsonFormsConfiguration} from './ducks/scidata/config'
 
+
 export default function configureStore( initialState ) {
     //const rootReducer = combineReducers( reducers );
 
@@ -16,15 +17,16 @@ export default function configureStore( initialState ) {
         rootReducer,
         initialState,
     );
+    console.log(store.getState())
        
     const data = {};
     
     store.dispatch(
       Actions.init(
-        data,
-        scidataJsonSchema,
-        scidataUiSchema,
-        scidataJsonFormsConfiguration
+          data,
+          scidataJsonSchema,
+          scidataUiSchema,
+          scidataJsonFormsConfiguration
       )
     )
