+++ src/views/components/scidataForm.css
@@ -1,19 +1,19 @@
-.App {
+.scidataForm {
   text-align: center;
   margin-bottom: 0.5em;
 }
 
-.App-header {
+.scidataForm-header {
   background-color: #222;
   height: 50px;
   padding: 20px;
   color: white;
 }
 
-.App-title {
+.scidataForm-title {
   font-size: 1.5em;
 }
 
-.App-intro {
+.scidataForm-intro {
   font-size: large;
 }
